+++ packages/@kagal-ed25519-secret/src/index.ts
@@ -27,5 +27,6 @@ export {
   type Bytes,
   decodeBase64,
   encodeBase64,
+  encodeKey,
   getRandom,
 } from './utils';